+++ sno/show.py
@@ -1,7 +1,5 @@
 import contextlib
-import json
 from datetime import datetime, timezone, timedelta
-from io import StringIO
 
 import click